==> notebooks/2-fetch-PSAW.ipynb <==
# Using PushShift as a PRAW wrapper to get even more data

import praw
import os
from dotenv import load_dotenv
load_dotenv()

reddit = praw.Reddit(client_id = os.environ.get('CLID'),
                    client_secret = os.environ.get('CLSC'),
                    username = os.environ.get('USER'),
                    password = os.environ.get('PW'),
                    user_agent = 'Cygni_praw')

from psaw import PushshiftAPI
api = PushshiftAPI(reddit)

top = api.search_submissions(subreddit='depression',limit=100000)
context = []
label = []
for sub in top:
    if len(sub.selftext) > 32:
        s = sub.selftext.split(sep='\n')
        for p in s:
            if len(p) > 32:
                context.append(p)
                label.append('depression')

data = {'text': context,
        'class': label}

import pandas as pd
df = pd.DataFrame(data)
df

df.to_csv('depression.csv')
