==> notebooks/3-df_join.ipynb <==
# Joining the dataset, train test splitting

import pandas as pd
ne = pd.read_csv('neither.csv')
su = pd.read_csv('suicide.csv')
de = pd.read_csv('depression.csv')

df = pd.concat([ne, su, de]).drop(columns=['Unnamed: 0'])

df

from sklearn.model_selection import train_test_split

train, test = train_test_split(df, test_size=0.2, random_state=69420)

train = train.reset_index(drop=True)
test = test.reset_index(drop=True)

train

test

train.to_csv('train.csv')
test.to_csv('test.csv')
