==> notebooks/4-cleaning.ipynb <==
# Data cleaning

import pandas as pd
train = pd.read_csv('train.csv')
test = pd.read_csv('test.csv')

import re
import string

emoji_pattern = re.compile("["
                        u"\U0001F600-\U0001F64F"  # emoticons
                        u"\U0001F300-\U0001F5FF"  # symbols & pictographs
                        u"\U0001F680-\U0001F6FF"  # transport & map symbols
                        u"\U0001F1E0-\U0001F1FF"  # flags (iOS)
                           "]+", flags=re.UNICODE)
                        
def clean_text(text):
    text = text.lower()
    text = re.sub('\[.*?\]', '', text)
    text = re.sub('[%s]' % re.escape(string.punctuation), '', text)
    text = re.sub('\w*\d\w*', '', text)
    text = re.sub('[‘’“”…]', '', text)
    text = re.sub('\n', '', text)
    text = emoji_pattern.sub(r'', text)
    return text


def clean(df):
    text_cleaned = []
    for text in df['text']:
        text_cleaned.append(clean_text(text))
    df['text'] = text_cleaned
    return df


train = clean(train)
test = clean(test)

test

train.to_csv('train.csv')
test.to_csv('test.csv')
