==> notebooks/6-BERTclassifier.ipynb <==
# PsychNLP : Training the model

> in this notebook, we will be fine tuning a BERT model to classify our text data that we have in the previous notebooks. The codes in this notebook will:

- Preprocess and tokenize our text data
- Turn labels into one-hot format
- Create a data input pipeline
- Construct last_hidden_state layers and other BERT layers
- Build the model
- Define training steps to fine-tune our BERT

Let's get started!

# Setting up

In this section we will:

- Mount google drive
- Import necessary libraries
- Import our data

from google.colab import drive
drive.mount('/content/drive')

!pip install -q transformers

import transformers
from transformers import BertModel, BertTokenizer
from torch.optim import AdamW
import torch
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from collections import defaultdict
from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix, classification_report
from transformers.optimization import get_linear_schedule_with_warmup

from torch import nn, optim
from torch.utils.data import Dataset, DataLoader
import torch.nn.functional as F

using a cuda device

device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
device

Importing training and testing data, dropping some unnecessary columns and dropping nan values

train = pd.read_csv('/content/drive/MyDrive/ML/train.csv').drop(columns=['Unnamed: 0','Unnamed: 0.1']).dropna()
test = pd.read_csv('/content/drive/MyDrive/ML/test.csv').drop(columns=['Unnamed: 0','Unnamed: 0.1']).dropna()

Some sanity checks to make sure the data is correct

train.head()

train.info()
test.info()

# Preprocessing dataset

In this section we will:

- Convert labels into one-hot encodings
- Define tokenizers to tokenize the text data

def class_to_onehot(df):
  OH = []
  for label in df['class']:
    if label =='neither':
      value = [1,0,0]
    elif label =='depression':
      value = [0,1,0]
    elif label=='suicide':
      value = [0,0,1]
    OH.append(value)
  df['class'] = OH
  return df


train = class_to_onehot(train)
test = class_to_onehot(test)

test

One-hot encodings are used in classification models

model_name = 'bert-base-cased'
tokenizer = BertTokenizer.from_pretrained(model_name)

Downloaded BERT base model and getting the BertTokenizer

class tokenize_ds(Dataset):
  def __init__(self, text, labels, tokenizer, max_len):
    self.text = text
    self.labels = labels
    self.tokenizer = tokenizer
    self.max_len = max_len
  
  def __len__(self):
    return len(self.text)
  
  def __getitem__(self, item):
    text = str(self.text[item])
    label = self.labels[item]

    encoding = self.tokenizer.encode_plus(
      text,
      add_special_tokens=True,
      max_length=self.max_len,
      return_token_type_ids=False,
      padding='max_length',
      truncation=True,
      return_attention_mask=True,
      return_tensors='pt',
    )

    return {
      'text': text,
      'input_ids': encoding['input_ids'].flatten(),
      'attention_mask': encoding['attention_mask'].flatten(),
      'labels': torch.tensor(label, dtype=torch.long)
    }


Final sanity checks

train.shape

# Data input pipeline

def CreateDataLoader(df, tokenizer, max_len, batch_size):
  DS = tokenize_ds(
      text = df['text'].to_numpy(),
      labels = df['class'].to_numpy(),
      tokenizer = tokenizer,
      max_len = max_len
  )
  return DataLoader(DS, batch_size = batch_size, num_workers = 2)

test, val = train_test_split(test, train_size = 0.5, random_state=69420)

batch_size = 8
max_len = 128

Traindata = CreateDataLoader(train, tokenizer, max_len, batch_size)
Valdata = CreateDataLoader(val, tokenizer, max_len, batch_size)
Testdata = CreateDataLoader(test, tokenizer, max_len, batch_size)

print(next(iter(Traindata))['input_ids'].shape)
print(next(iter(Traindata))['attention_mask'].shape)
print(next(iter(Traindata))['labels'].shape)

next(iter(Traindata)).keys()

# Constructing model

bert_model = BertModel.from_pretrained(model_name, return_dict=False)

train.keys()

encoding = tokenizer.encode_plus(
  str(train['text'][0]),
  max_length=128,
  add_special_tokens=True,
  return_token_type_ids=False,
  padding='max_length',
  truncation=True,
  return_attention_mask=True,
  return_tensors='pt', 
)


outputs = bert_model(
  input_ids=encoding['input_ids'], 
  attention_mask=encoding['attention_mask']
)

last_hidden_state = outputs[0]
pooled_output = outputs[1]

last_hidden_state.shape

pooled_output.shape

bert_model.config.hidden_size

# Model Building

class TextClassifier(nn.Module):
  def __init__(self, n_classes):
    super(TextClassifier, self).__init__()
    self.bert = BertModel.from_pretrained(model_name, return_dict=False)
    self.drop = nn.Dropout(p=0.3)
    self.L1 = nn.Linear(self.bert.config.hidden_size, n_classes)
    self.out = nn.Softmax(dim=1)
    

  def forward(self, input_ids, attention_mask):
    _, pooled_output = self.bert(
        input_ids = input_ids,
        attention_mask = attention_mask
    )
    output = self.drop(pooled_output)
    output = self.L1(output)
    output = self.out(output)
    return output

model = TextClassifier(3)
model.bert.requires_grad_(False)
model = model.to(device)

data = next(iter(Traindata))
input_ids = data['input_ids'].to(device)
attention_mask = data['attention_mask'].to(device)

input_ids.shape

print(model(input_ids, attention_mask))

# Training Model

epochs = 10

optim = AdamW(model.parameters(), lr = 1e-3)
steps = len(Traindata) * epochs

scheduler = get_linear_schedule_with_warmup(
    optim,
    num_warmup_steps=0,
    num_training_steps=steps
)

loss = nn.CrossEntropyLoss().to(device)

def train_epoch(
    model,
    data_loader,
    loss,
    optim,
    device,
    scheduler,
    n_examples
  ):
  model = model.train()

  losses = []
  correct_predictions = 0.0
  p = 0
  for d in data_loader:
    if p % 100 == 0:
      print(f'p = {p} out of {len(data_loader)}')
    input_ids = d['input_ids'].to(device)
    attention_mask = d['attention_mask'].to(device)
    labels = d['labels'].type(torch.FloatTensor).to(device)
    optim.zero_grad()
    outputs = model(
      input_ids = input_ids,
      attention_mask = attention_mask
      )
    preds = torch.argmax(outputs, dim = 1)
    #print(torch.argmax(labels.float(), dim=1).long().cpu().numpy())
    #print(preds.cpu().numpy())
    loss_value = loss(outputs, labels)
    #print(np.count_nonzero((preds.cpu().numpy() == torch.argmax(labels.float(), dim=1).long().cpu().numpy())))
    correct_predictions += np.count_nonzero((preds.cpu().numpy() == torch.argmax(labels.float(), dim=1).long().cpu().numpy()))
    losses.append(loss_value.item())

    loss_value.backward()
    nn.utils.clip_grad_norm_(model.parameters(), max_norm=1.0)
    optim.step()
    scheduler.step()
    
    p+=1

  return correct_predictions/ n_examples, np.mean(losses)

def eval_model(
    model,
    data_loader,
    loss,
    device,
    n_examples
):
  model = model.eval()
  losses = []
  correct_predictions = 0
  p = 0

  with torch.no_grad():
    for d in data_loader:
      if p % 100 == 0:
        print(f'p = {p} out of {len(data_loader)}')
      input_ids = d['input_ids'].to(device)
      attention_mask = d['attention_mask'].to(device)
      labels = d['labels'].type(torch.FloatTensor).to(device)

      outputs = model(
        input_ids = input_ids,
        attention_mask = attention_mask
        )
      preds = torch.argmax(outputs, dim = 1)
      #print(preds)
      loss_value = loss(outputs, labels)

      correct_predictions += np.count_nonzero((preds.cpu().numpy() == torch.argmax(labels.float(), dim=1).long().cpu().numpy()))
      losses.append(loss_value.item())
      p +=1
  return float(correct_predictions) / n_examples, np.mean(losses)

def training_step(epochs):
  %%time

  history = defaultdict(list)
  best_acc = 0
  for epoch in range(epochs):
    print(f'Epoch {epoch + 1} / {epochs}')

    train_acc, train_loss = train_epoch(
        model,
        Traindata,
        loss,
        optim,
        device,
        scheduler,
        len(train)
    )
    print(f'Training loss = {train_loss}, accuracy = {train_acc}')
    

    val_acc, val_loss = eval_model(
        model,
        Valdata,
        loss,
        device,
        len(val)
    )
    print(f'Validation loss = {val_loss}, accuracy = {val_acc}')

    history['train_acc'].append(train_acc)
    history['train_loss'].append(train_loss)
    history['val_acc'].append(val_acc)
    history['val_loss'].append(val_loss)

    if val_acc > best_acc:
      torch.save(model.state_dict(), 'best_model_state.bin')
      best_acc = val_acc

# Actual model training

training_step(epochs)
